modified:   common/src/main/java/com/scut/common/dto/request/BoardCommentParam.java 	modified:   common/src/main/java/com/scut/common/dto/response/BoardCommentDto.java 	modified:   common/src/main/java/com/scut/common/dto/response/GameDetailsDto.java 	modified:   common/src/main/java/com/scut/common/dto/response/GameForBoardDto.java 	modified:   common/src/main/java/com/scut/common/dto/response/SearchGameListDto.java

+++ common/src/main/java/com/scut/common/dto/response/GameDetailsDto.java
@@ -10,15 +10,15 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @ApiModel(description = "游戏信息")
 public class GameDetailsDto {
-    int id; //游戏id
+    long id; //游戏id
     String name; //游戏名称
     String cover; //封面地址
     String logo; //logo地址
-    int issuedTime; //游戏上架时间
+    long issuedTime; //游戏上架时间
     String downloadLink; //下载链接
     String classify; //分类
     float score; //评分
-    int scoreCount; //评分数量
+    long scoreCount; //评分数量
     String totalScore; //评分总和
     String description; //简要描述
 }
